Referrers
phew! what a long commit. It's basically a new component (table) which
allows creating a simple table, plus all the internals for the parsing
and querying of referrers the backend. I've also added a bar chart
to the mix, and thrown in some refactoring here and there for good
measure. I've also added a couple of tests.

added 'unkown' sanitation to data