Referrer URL Table
So i've added a referrer table that has the links to the articles.
Unfortunately there is no straightforward way of loading twitter urls
from t.co urls, but the rest is there - we now have a list of referral
urls that are links which actually open. How cool is that? 💪

only internal!

fixing my botched rebase