Get URL's for internal Referrers
I've added functionality to query the `<title>` tags extracted from our
internal referrer urls by loading the pages and running a regex on the
response from the server. The titles for the URL's are cached on redis
so as to avoid hammering servers.

This means that if you don't have redis installed on your dev machine
you might want to install it to avoid scraping websites too intensely.

Some websites (ft.com) throw errors and thus we show the urls instead

fix for no redis