Add a scroll To Top functionality on route entering
when we enter a route, we will scroll the page to the top in order to
avoid journalists from getting lost in the page, saving them valuable
seconds of scrolling on their otherwise busy lives.

Also fixed the constructor of the `SectionModifier.js` where
@radiodario drunkenly changed the initial state into a `setState` call,
inadvertently creating a no-op and being a general muppet.